services/puzzle-agent/requirements.txt
# Core Dependencies for scavenger.RiddleAgent
Flask==3.1.1
pymongo==4.13.2
python-dotenv==1.1.1
openai==1.97.1
lmstudio==1.4.1

# Additional dependencies that may be pulled in
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0

# HTTP and networking
httpx==0.28.1
httpcore==1.0.9
certifi==2025.7.14
idna==3.10
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0

# Data processing
pydantic==2.11.7
pydantic_core==2.33.2
annotated-types==0.7.0
typing_extensions==4.14.1
